Add support for MULTIARCH_SKIP_DIRS.{bin,lib}.
The primary use-case for these is perl, where the multiarch
directories are under PERL5_INSTALLVENDORARCH rather than "lib".

To handle packages which build both shared libraries and perl
modules, for example net/net-snmp, we need to ensure that "lib" is
correctly expanded, while "lib/perl5" is left alone, and in this
case we can add "lib/perl5" to MULTIARCH_SKIP_DIRS.lib.